feat: 연산자 추출 실패하는 메서드
